a lot of changes have happend from this point
