+++ client/src/app.ts
@@ -17,6 +17,8 @@ import COLORS from './colors'
 //import { MapGenHelper } from './mapGenHelper'
 
 import SETTINGS from './gameSettings'
+import { handleInput } from './handleINput'
+import { renderToGrid } from './renderToGrid'
 // sizing
 
 
@@ -51,8 +53,6 @@ if(!seedStr){
     RANDOM.seed(seedStr)
 }
 
-
-
 const player: Entity = new Entity(ID_MANAGER.next(), 3,4, '@', COLORS.player)
 const npc: Entity = new Entity(ID_MANAGER.next(), 3,4, '@', COLORS.npc)
 const entities: Entity[] = [player,npc]
@@ -66,7 +66,6 @@ const km = new KeyboardMonitor().attach(document)
 const mm = new MouseMonitor().attach(canvas)
 
 const renderer = new CanvasRenderer()
-
 const cameraFrame = Rect.make(0,0, CAMERA_WIDTH, CAMERA_HEIGHT) // CAMERA IS IN WORLD CELLS, NOT ABSOLUTE UNITS
 const renderGrid = new Grid<IRenderCell>(cameraFrame.width, cameraFrame.height)
 renderGrid.setEach((cell: any, index: number, x: number, y: number): IRenderCell => {
@@ -93,9 +92,6 @@ const rooms: IRect[] = []
 mapGenerator1(tileGrid, rooms)
 //mapGenerator2(tileGrid, rooms)
 
-// const cameraFrame = Rect.make(0, 0, 10, 10)
-
-// SET Entity Locations
 {
     const pcenter = Rect.center(rooms[0])
     Point.set(player, pcenter.x, pcenter.y) 
@@ -107,79 +103,6 @@ mapGenerator1(tileGrid, rooms)
     cameraFrame.y = Math.floor(player.y - cameraFrame.height/2)
     fovGrid.x = cameraFrame.x
     fovGrid.y = cameraFrame.y
-}
-
-// We need to move this into it's own section, and potentially allow substituting it out
-const renderToGrid = (tileGrid: Grid<Tile>, fovGrid: Grid<FOVCell>, entities: Entity[], renderGrid: Grid<IRenderCell>): void => {
-    // renderGrid is in SCREEN coordinates, and will have it's XY ignored for our purposes
-    // tileGrid, entities and fovGrid are in WORLD coordinates and will have their XY and y ignored for now
-    
-    // camera and renderGrid both start at the topLeft in terms of aligning the two
-    // I almost feel like we should tighten the coupling, but oh well, let's actually do it
-    const screenP = Point.make(0,0)
-    const worldP = Point.make(0,0)
-    for(let relCameraY = 0; relCameraY < cameraFrame.height; relCameraY++){
-        for(let relCameraX = 0; relCameraX < cameraFrame.width; relCameraX++){
-            // this maps to the renderGrid and the cameras
-            Point.set(screenP, relCameraX, relCameraY)
-            // this maps to the tileGrid, fovGrid, and entities
-            Point.set(worldP, screenP.x + cameraFrame.x, screenP.y + cameraFrame.y)
-            
-            // Let's do some bounds checking
-            // we know these poitns are always in cameras, so never need to check that
-            // we are going to offload the case of checking if the renderGrid can fit the cameraSize for now
-            // so for now just check if we are inside of the tileGrid, which currently should be the same dimensions as the fovGrid, though that is bound to change as we try and optimize things
-            // soooooo, let's check if we are inside the tileGrid
-            const renderCell  = renderGrid.getP(screenP)
-            if(tileGrid.inBoundsXY(worldP.x, worldP.y)){
-                const tile = tileGrid.getP(worldP)
-                const fovCell = fovGrid.getP(screenP)
-                renderCell.character = ''
-                if(tile.explored){
-                    if(!fovCell.visible){
-                        if(tile.blockMove){
-                            renderCell.backColor = COLORS.dark_wall
-                        } else if(tile.blockSight){ // it's a door
-                            renderCell.backColor = COLORS.dark_door
-                        } else {
-                            renderCell.backColor = COLORS.dark_ground
-                        }
-                    } else {
-                        if(tile.blockMove){
-                            renderCell.backColor = COLORS.light_wall
-                        } else if(tile.blockSight){ // it's a door
-                            renderCell.backColor = COLORS.light_door
-                        }
-                        else {
-                            renderCell.backColor = COLORS.light_ground
-                        }
-                    }
-                } else {
-                    renderCell.backColor = COLORS.black
-                
-                }
-            } else {
-                renderCell.character = ''
-                renderCell.foreColor = COLORS.black
-                renderCell.backColor = COLORS.black
-            }
-        }
-    }
-
-    entities.forEach((entity: Entity): void => {
-        Point.set(screenP, entity.x - cameraFrame.x, entity.y - cameraFrame.y)
-        // if this is inside of the camera bounds
-        if(cameraFrame.x <= entity.x && entity.x <= cameraFrame.x + cameraFrame.width - 1 &&
-            cameraFrame.y <= entity.y && entity.y <= cameraFrame.y + cameraFrame.height - 1){
-            const rCell: IRenderCell = renderGrid.getP(screenP)
-            const fovCell: FOVCell = fovGrid.getP(screenP)
-            if(fovCell.visible){
-                rCell.foreColor = entity.color
-                rCell.character = entity.character
-            }
-        }
-        
-    })
 }
 
 // clean this up
@@ -204,24 +127,7 @@ loadImage('assets/out.png').then((image: any): void => {
         // we need to translate this into a "move" type action
         // we'll just do the move section
         
-        if(km.getKeyState('ArrowLeft').isDown && (km.getKeyState('ArrowLeft').halfSteps > 0 || km.getKeyState('z').isDown)){
-            PUBSUB.publish('move', {id: player.id, delta: Point.make(-1, 0)})
-        }
-        if(km.getKeyState('ArrowRight').isDown && (km.getKeyState('ArrowRight').halfSteps > 0 || km.getKeyState('z').isDown)){
-            PUBSUB.publish('move', {id: player.id, delta: Point.make(1, 0)})
-        }
-        if(km.getKeyState('ArrowUp').isDown && (km.getKeyState('ArrowUp').halfSteps > 0 || km.getKeyState('z').isDown)){
-            PUBSUB.publish('move', {id: player.id, delta: Point.make(0, -1)})
-        }
-        if(km.getKeyState('ArrowDown').isDown && (km.getKeyState('ArrowDown').halfSteps > 0 || km.getKeyState('z').isDown)){
-            PUBSUB.publish('move', {id: player.id, delta: Point.make(0, 1)})
-        }
-        // quick number generator
-        if(!km.getKeyState('n').isDown && (km.getKeyState('n').halfSteps > 0)){
-            const seed = btoa(new Date().toString())
-            const newurl = window.location.protocol + '//' + window.location.host + window.location.pathname + '?seed=' + seed
-            window.location.href = newurl
-        }
+        handleInput(km, player)
 
         // process moves
         // eventually we'll componentize this better
@@ -232,30 +138,31 @@ loadImage('assets/out.png').then((image: any): void => {
             if(!mover){
                 return
             }
-            if(tileGrid.inBoundsXY(player.x + move.x, player.y + move.y)){
+            if(tileGrid.inBoundsXY(mover.x + move.x, mover.y + move.y)){
                 const tile = tileGrid.getXY(player.x + move.x, player.y + move.y)
                 if(!tile.blockMove){
                     mover.move(move.x, move.y)
                     if(mover.id === player.id){
                         fovRecompute = true
-                        PUBSUB.publish('moved', msg) // only publish move if the entity actually completed the move
                     }
+                    PUBSUB.publish('moved', msg) // only publish move if the entity actually completed the move
                 }
             }            
         })
         MoveProcessor.moves.length = 0
-        if(fovRecompute){
-            calculateFOV(fovGrid, tileGrid, player, FOV_RADIUS)
-        }
-
+        
         // determine camera position
         cameraFrame.x = Math.floor(player.x - cameraFrame.width/2)
         cameraFrame.y = Math.floor(player.y - cameraFrame.height/2)
         fovGrid.x = cameraFrame.x
         fovGrid.y = cameraFrame.y
 
+        if(fovRecompute){
+            calculateFOV(fovGrid, tileGrid, player, FOV_RADIUS)
+        }
+
         // we might move all of this into some offscreen 
-        renderToGrid(tileGrid, fovGrid, entities, renderGrid)
+        renderToGrid(tileGrid, fovGrid, entities, renderGrid, cameraFrame)
         renderer.render(renderGrid)
 
         // ctx.drawImage(image, 0, 500)
